Measure 3: move the broker registry to datapipeline
EmbeddedBrokerProcess held two roles in one class, with the seam already
marked in its own source: a thin process wrapper, and a static registry
plus the entire Artemis configuration. Every caller of that static half
is in datapipeline; node never used it. Unlike the runtime edge this is
not observation but infrastructure on the operating path, so it was
simply in the wrong package.

The static half becomes EmbeddedBrokerRegistry in
datapipeline.resources.broker, a sibling of the two packages that consume
it. Its public surface is exactly the previous static API — nothing added.
configureLogging becomes private: ensureStarted is its only caller.

The wrapper shrinks from 566 to 116 lines and keeps only the lifecycle.
stop() asked the private brokerRegistry field whether a broker was
running; it now asks isBrokerStarted, which already existed for that
purpose. No Artemis import remains in node. The process keeps its name
and package, so reference.conf is untouched.

ArtemisTopicResource also read DEFAULT_MAX_SIZE_BYTES, a constant the
compiler inlines — no bytecode reference, so the rule never reported it.
Repointing only what the rule names would have left that import behind
with the rule still green.

shouldParseInVmServerId moves to EmbeddedBrokerRegistryTest and becomes a
unit test: it is a pure function and needs no running broker. The four
lifecycle tests stay with the process.

The TopicCleaner JavaDoc pointed at configureLogging to explain a
null-check. That link was never resolvable across packages and is now
prose, which keeps the reasoning without the broken reference.

datapipelineDependsOnCompilerAndRuntimeOnly is green; cycles drop from
five to three, all remaining ones between compiler and its neighbours.
